修改获取请求时间类型错误
$contextData['requestTime']是一个浮点型

+++ src/framework/src/Log/Logger.php
@@ -485,7 +485,7 @@ class Logger extends \Monolog\Logger
      *
      * @return int
      */
-    private function getRequestTime(): int
+    private function getRequestTime(): float
     {
         $contextData = RequestContext::getContextData();
